replaced testimonials and added migrations
